requirements.txt
sphinx==8.1.3
sphinx-needs>=1.0.0
sphinx-simplepdf==1.6.0
sphinxcontrib-plantuml
bitfield-extended
drawsvg
setuptools
